gnome-terminal-2.22 in portage with some changes:
* sed out G_DISABLE_DEPRECATED and co in this version too for future-proofing. Upstream bug #523737 filed on that.
* Add a patch from upstream SVN branch to fix broken "Tabs->Detach tab" menu option

Note that gnome-terminal-2.22 has a regression in custom backgrounds image selection - there is no image preview, while there was one in 2.18.
This code was ported to GIO and for some reason I'm getting a NULL returned from g_file_query_info in create_preview_pixbuf in file profile_editor.c
with GError set to G_IO_ERROR "Operation not supported", iirc. Image preview fails for several people on IRC as well.
Not considering this a blocker for adding to portage p.masked, nor unmasking it to ~arch. I'm sure no-one would have even noticed this small regression in this eye-killer feature if I
hadn't tested all the functionality that was changed in gnome-terminal-2.22 compared to 2.18 based on ebuild unpacked directory differences and ChangeLog.