Kill empty directories that have just a ChangeLog and other staleness in it.
Please do that always when there is no ebuild in it anymore, because

a) It allows for easy seeing what's in the overlay with find /path/to/overlay -mindepth 2 -maxdepth 2 -type d
b) We do not use ChangeLog files, but git log
c) The full history is always available even if the directory is gone. If the directory reappears then you can just git log /path/to/pkgdir, and if
it's gone, you simply have to be explicit that it's a dir/file, not a branch name (it decides based on a stat() it seems). Like this:

git log -- dev-libs/glib  # shows the full history of glib, even if that directory doesn't exist currently