gnome-base/gail: Change the PDEPEND on gtk+ to a RDEPEND
This allows portage-2.2_rc* to automatically steal the gail libraries
ownership from gail thanks to gtk+ having the gail blocker, and it
being ensured to upgrade before gail so the magic works.
ABCD from #gentoo-portage helped to figure this one out and he thinks
that's also how the ss/com_err -> e2fsprogs-libs mess got solved.
With it having been PDEPEND you get a file collision if preserved-libs
is enabled, but not with RDEPEND.
emerge -1va gtk+ will not work if gail is also installed, but
a regular @world update upgrades both and it works fine then,
as does emerge -1va gail