x11-libs/gtk+: Bump to 2.14.3 and make it be closer to what I'm about to put to portage
The version in portage will not have the DirectFB support because without lots of
checking, warning and so on it might cause everything to break for some users and
we don't need to deal with that right now. Hence this version will be kept in here
as well for the time being.

Changes to the ebuild include:
* Use jpeg2k _global_ USE flag for jpeg2000 gdk-pixbuf module building
* depend on libXrandr-1.2 and newer, because that's what the configure.in checks now for
having better multihead support, and we only have 1.2 versions in tree anyhow
* remove the openoffice workaround patch - shouldn't be necessary anymore
* call elibtoolize in the end as there is no eautoreconf anymore and portage version told
to add that back if eautoreconf isn't needed anymore
* Fix indent to be what it was in portage. No need to graciously do extra whitespacing so
that it would line up for someone, while for someone with a different tab space setting it
is worse then. One tab for each indent and end of story.

Overview of Changes from GTK+ 2.14.2 to 2.14.3
==============================================

* Revert problematic GtkAdjustment changes

* Bugs fixed:
 552837 mem leak in gtkimmulticontext
 553000 incorrect i18n header in gtkfilesystem.c
 553135 eog crash: assertion failed
 552545 leaks GpImage

* Updated translations