libgnomeui in portage without jpeg USE flag and gtk+ dep increased to 2.14
--with{,out}-jpeg has never had any effect to libgnomeui, as it probably uses gdk-pixbuf, and in 2.24 this was removed
completely. Not sure why not in the overlay ebuild... *grin*
gtk+ dep increased from 2.12 to 2.14 as the only big change in libgnomeui-2.24 is the removal of all the VFS backends,
because they can't be built against gtk+-2.14 and gtk provides this functionality itself now, so the dep increase
ensures the user will still have VFS functionality.