gnome-extra/gtkhtml: Some fixes to the ebuild
Note that this version does not work for GNOME-2.22, because evolution-2.22 uses the bonobo editor and with
gtkhtml-3.24 it doesn't exist by default, so it can't load it and e-mail composing is not possible.
However we could pass --with-bonobo-editor to still build it for backwards compatibility for now (putting
libbonobo{,ui} and orbit deps back in that case) and then it _should_ work (untested) with evo-2.22 too -
undecided about the enabling, needs checking with other gtkhtml users.

Ebuild changes:

* Kill USE=static, but instead of always building it, always disable it; there are no users of it.
* Remove the libbonobo(ui)/orbit dependencies - should not be necessary unless we pass --with-bonobo-editor, but hard to test because libgnome check requires the bonobo pc file.
* Increase iso-codes dep to >=0.49 - iso-codes doesn't get enabled unless this dep is satisfied because of some kind of licensing issues
* Add gconf dependency, used directly, also listed in GTKHTML_MODULES variable that is passed to pkgconfig
* Remove --enable-file-chooser passing. The switch has been removed ever since gtkhtml-3.15.6 and since then it is just always enabled