gconf: Really re-add policykit support.
Gilles needs a ebuild/bash syntax highlighting editor.

PolicyKit support actually NOT tested by me of course, but it's
supposed to allow syncing some keys to be system default ones,
such as desktop wallpaper, so gdm will see that and show your
nice gentoo wallpaper to everyone - probably wildly inappropriate
with the most important part covered by the login box.
But you theoretically get nice fade-in into the desktop if the
user logging in happens to be you.
No idea what actually does that key sync to system (and the policykit
rules for that, or other such options).