scripts/eshowkw: Add. This version doesn't spam about prefix arches.
Pass -a as first argument to this version of eshowkw to see all arches,
including prefix arches. It relies heavily on the comment in the arch.list
file in portage tree and wrongly assumes that comment won't change.
But good enough for now to reduce the spammy output (when the comment changes,
we simply see all prefix arches by default again).