udisks+gnome-disk-utility: Add a masked udisks 1.0.0 ebuild, and bump gnome-disk-utility
This really needs to be cleaned up.  It installs an avahi service file,
a bash-completion file.  Those are the ones off the top of my head.
Probably more.  It also requires a newer lvm2 than we have in portage,
but we dep on lvm2 anyway for device-mapper support since udev can't be
installed with parted[device-mapper].