evolution & friends: 3.0.1 → 3.0.2 + out-of-tree linking fix for eds
* mail-client/evolution: 3.0.1 → 3.0.2, add a few patches from git
gnome-3-0 branch, update the default browser message for GNOME3
(both in 3.0.2 and 9999).

* gnome-extra/evolution-exchange: 3.0.1 → 3.0.2, remove obsolete patch.

* gnome-extra/evolution-mapi: 3.0.1 → 3.0.2, drop all keywords until
libmapi-0.9 can be compiled on Gentoo without the use of black magick
(e.g. see bug 355935, and that's just the tip of the iceberg).

* gnome-extra/evolution-data-server: 3.0.1 → 3.0.2.1, plus fix the
out-of-tree linking issue. A long explanation follows.

evolution-data-server tries to link to system libraries rather than to
in-tree ones on make install (see https://bugzilla.gnome.org/show_bug.cgi?id=641239)
making it impossible to e.g. cleanly upgrade from gnome2 to gnome3. After
spending hours banging my head against the abomination known as libtool,
I see no clean and simple way of fixing this: no matter what I tried,
the chain of.la dependencies resolved into an -L/usr/lib64 somewhere in
front of -lcamel-1.2 or -lsome-other-in-tree-lib.

Therefore, as an alternative to rewriting eds's build system, I propose
adding a perl script (fix_relink_command.pl) that runs before
gnome2_src_install and fixes up all the relink_command lines in the .la files
in the build tree to make libtool preferentially relink to in-tree libraries.

I have tested it, and works well, allowing a clean upgrade from
evolution-data-server-2.32.2-r1 to 3.0.2.1.

Why perl? Because the kind of text mangling that needs to be performed is
really not a good fit for tools like sed, so a shell script with equivalent
functionality would be longer, less readable, and probably not portable
to non-GNU systems.

Signed-off-by: Nirbheek Chauhan <nirbheek@gentoo.org>