media-sound/rhythmbox: 0.13.3 was moved to gx86
Punt rb 0.13.3 and stale patches (in gx86 since March). Update the live
ebuild with changes from the in-gx86 ebuild (new udev flags, pkg_setup
changes) and otherwise update dependencies (runtime and build-time deps
on pygobject and libpeas; the coherence plugin imports gconf and pygtk
for some strange reason) to reflect the current state of the tree.

Other than the vala plugin (which does not build at the moment), the
live ebuild appears to compile and work.

Signed-off-by: Nirbheek Chauhan <nirbheek@gentoo.org>