gnome-extra/gucharmap: slot gtk3 version to 2.90 (bug #362035)
gnome-extra/gucharmap installs a library that is used by several
packages to display a unicode character picker tool. Some of these
packages - such as abiword, bluefish, and libtomoe-gtk - are unlikely to
be ported to gtk3 in the immediate future. It is therefore desirable to
be able to install the gtk2 version of the gucharmap library in
parallel with the gtk3-based one.

The solution is to slot the gtk2 version to "0" and the gtk3 version to
"2.90". Both versions can be built from the same 3.0.x source via the
appropriate --with-gtk configure option. If we ensure that only one of
the slots (2.90, naturally) installs the gucharmap program and
associated .desktop file and schemas, and rename slot 2.90's gettext
domain, there will be no file collisions.

This partially addresses bug #362035.

v2:
Mgorny on IRC pointed out that my original slotmove command would have
moved gucharmap-3.0.1-r200 to slot 2.90 the next time updates/2Q-2011
changed. So instead, explicitly list all versions of gucharmap for gtk3
that have ever been in the overlay.