app-text/gtranslator: fix charmap plugin
In an IRC conversation with nirbheek, it became apparent that the
gtranslator ebuild does not take into account the program's charmap
plugin, written in python. As a result, make the following changes
to the ebuild:
* add correct dependencies (gucharmap:2.90[introspection] etc.) for
the charmap plugin when USE=gnome, and python_mod_optimize it.
* don't install the plugin when USE=-gnome.
* drop the separate introspection flag from IUSE; the only possible
reason for introspecting gtranslator is to run the charmap plugin.