dev-cpp/libgdamm: 4.99.0.1 → 4.99.3
Version bump. Switch to EAPI 4 to get the xz tarball. Add a berkdb USE flag
and RDEPEND on libgda:5[berkdb=] due to an automagic dependency.

Note that configure in the tarball has a libgda-4.99.3 requirement. This is
obviously a typo since:
a. libgdamm-4.99.3 was released 13 hours after libgda-4.99.2, i.e. at the
point in time when there was no difference between libgda git master (the
future 4.99.2) and the libgda 4.99.2 release; and
b. the main change in libgdamm-4.99.3 is suport for new null semantics that
were introduced in the libgda-4.99.2 API.