gnome-base/gnome-desktop: install pnp.ids file
A number of Gnome 3.1 packages require the pnp.ids file that on Gentoo
is provided by gnome-desktop; ensure that it is installed correctly by
*not* using --with-pnp-ids-path configure argument (if it's set to
anything, pnp.ids will not be installed).

Since the set of installed files changed, bump the revision.

Also, update the live ebuild with changes from 3.0.x, update the
intltool dependency, and get the pnp.ids file from Fedora's hwdata git
(requiring an additional call to git-2_src_unpack), because the file is
only included in the gnome-desktop tarballs, not in gnome-desktop git.