app-portage/packagekit: split and moved to app-admin in gx86
packagekit is now in gx86, in app-admin, and split into 4 packages:
packagekit-base, packagekit-gtk, packagekit-qt, and packagekit (the
meta package). Many thanks to lxnay and the sabayon people for doing
this work.

Unfortunately, there is no way to provide a smooth upgrade path from
monolithic packagekit in the overlay to the split one in gx86,
particularly when both the overlay and gx86 contain the same version of
packagekit. The best one can do is to add an app-admin/packagekit-base
ebuild to the overlay that hard-blocks the old app-portage/packagekit.

In addition, all the packagekit dependencies in the overlay have had to
be updated. Most of them (with the notable exception of
gnome-packagekit) only need packagekit-base.