gnome-extra/gnome-color-manager: 3.0.0 → 3.1.2
Bump to gnome-3.1.x version of gnome-color-manager and add a live
ebuild. The dependencies have changed significantly (much of g-c-m's
functionality has moved to colord and gnome-control-center). Ensure that
gnome-control-center-3.1.x blocks g-c-m-3.0.x due to file collisions.
Move the block in gnome-settings-daemon to rdepends so as not to prevent
people from building binary packages with --buildpkgonly.