dev-libs/gjs: switch from spidermonkey to xulrunner USE flag
Control the choice of javascript engine with the "xulrunner" USE flag
instead of "spidermonkey". First, xulrunner is already a global USE
flag; second, spidermonkey is preferred by upstream; third, apparently
xulrunner in gentoo will be going away at some point in the future.

Since spidermonkey-1.8.5 is currently masked (and breaks lots of stuff),
make the xulrunner flag default + for now. When this ebuild is moved to
gx86, the xulrunner flag should be forced on with package.use.force until
the new spidermonkey is unmasked.