dev-libs/gjs: add gjs-0.7.14-r1 with optional spidermonkey support
Port the changes from gnome-next branch to enable gjs to build with
either xulrunner-2 or spidermonkey-1.8.5. Required creating a patch
that backports the relevant improvements that exist in upstream git
master's configure. Builds and passes tests with either choice of
javascript engine.

Note: the commits taken from upstream git keep on rewriting the same
region of configure script, removing or altering the effect of previous
changes. Therefore, the patch is provided in squashed form, both to use
less space, and to make it much easier for an outsider to see what it
has actually changed in the 0.7.14 code.