dev-python/pygobject: add 2.90.1 for gobject-introspection-1.29.x support
Add a version of pygobject that supports gobject-introspection-1.29.x,
since many (all?) python applications using gi from pygobject-2.28. fail
at runtime when gobject-introspection-1.29.x is installed. (Note: this
behavior may warrant a blocker, but need to think hard about whether/how
to implement such a blocker while leaving an upgrade path clear.)

Upstream recommends installing pygobject-2.28.x *without* introspection
in parallel with pygobject-2.90.1 (which *only* uses introspection); see
https://bugzilla.gnome.org/show_bug.cgi?id=642048#c9 and
http://www.daa.com.au/pipermail/pygtk/2011-August/019920.html

Therefore, we add pygobject-2.90.1 as slot 3, and make the introspection
flag in slot 2 pull in slot 3 as PDEPEND. Note that getting rid of the
introspection USE flag in slot 2 cannot be done in the short term due to
the number of packages that check for it.

Unfortunately, one side effect of this is that we have to RESTRICT test
in slot 2, since tests don't run without introspection. Need to think of
how to work around this issue.

And in 2.90.1, tests are also RESTRICTed for now because they require
gobject-introspection-1.29.16 (not released yet).