gnome-base/gnome-menus: add 3.1.5
Add gnome-3.1.x version of gnome-menus. Per upstream recommendation, put
it in a different slot (3). However, only the library is parallel-installable;
the gmenu-simple-editor program and directory files will collide, so need
to add 3.0.1-r50 that only installs the library (and optionally, the
python module and introspection data), and thus can be installed in
parallel with slot 3. This cut-down slot 0 pdepends on slot 3 to make
sure that people installing gnome-menus of whatever slot still get the
menu directory files.