dev-util/anjuta: 3.1.5 → 3.1.90
Version bump with numerous changes. Now uses libgda:5 and vala-0.13.3.

Forced to restrict tests because there is no way to pass them on Gentoo:
1. tests require plugins/am-project/tests/anjuta.lst to run, but this
file is not distributed with tarballs (see https://bugzilla.gnome.org/show_bug.cgi?id=657589);
2. in any case, anjuta.lst in upstream git is currently out of date (needs
to be updated for vala:0.14 and libgda:5, for example);
3. and, as far as I can tell, the tests will only pass when anjuta is
built with a specific set of USE flags.

Need to think of a way of handling this...