gnome-extra/gnome-user-docs: add 3.2.0.1
Update the license to CC-Attribution-3.0.
Update the dependencies (the build system now uses itstool and xmllint,
and scrollkeeper, mentioned in the spec file, is not actually used
enywhere; the documentation is now in mallard format instead of
docbook).
Drop the test USE flag since "make check" does not need any additional
tools besides xmllint and itstool.