app-misc/tracker: rename firefox USE flag, support firefox-bin, thunderbird-bin
Rename firefox USE flag to firefox-bookmarks; the firefox USE flag was
being automatically enabled in desktop profiles, which was not a good
idea because it has noticeable side effects on browser performance (i.e.
occasional stutters), and some people have reported that it even makes
firefox crash.

Also, support firefox-bin and thunderbird-bin by setting up the
appropriate symlinks. Instead of the former configure sedding hack,
have the ebuild create wrapper scripts that are the equivalent of
max("firefox --version", "firefox-bin --version") and similarly for
thunderbird. This also allows us to avoid an eautoreconf.