app-misc/tracker: put a stake through the foul heart of access violations (#385495)
It looks like (a) overriding AC_PATH_PROG requires a full path (not just
a relative one), and (b) AC_PATH_PROG actually tries to execute the
programs that it finds, so we need to generate the firefox and thunderbird
version wrapper scripts and pass them to configure, even if we are not
building firefox or thunderbird support.
Thanks to Stefan Zwanenburg <stefanhetzwaantje@gmail.com> for reporting
in bug #385495.