gnome-extra/gnome-user-docs: do not rebuild translations
Do not rebuild translations, they come with the tarball, and rebuilding
them takes far too long. Since translations are rebuild on "make all",
this simply amounts to skipping src_compile(). Thanks to Mart Raudsepp
for pointing out that the default behavior of the package could not
possibly be right.

Also, since we are not rebuilding translations, we can drop some
dependencies.

Now, installing gnome-user-docs with all translations takes seconds
instead of hours!