net-misc/networkmanager: build against libnl3 (#388609)
Build against libnl3 now that it has been unmasked, since nm
automagically detects the libnl version. Add some upstream patches to
allow libnl3 to be detected correctly. Add a patch to consistently use
linux/if.h instead of net/if.h (a good idea in general, and not doing so
leads to build failure with libnl3). Fixes bug #388609, thanks to
Yellowhat <yellowhat46@gmail.com> for reporting.