gnome-base/gdm: 3.5.92.1 → 3.6.0 with daemonization
Rewrite daemonization patch so it really works (and use libdaemon to get
it right). Ensure that "/etc/init.d/xdm stop" works by starting a new
process group if gdm is in daemon mode, and killing it in the signal
handler callback. Rewrite pam.d files to hopefully address some of the
problems that systemd users have been having (see bug #435042). And ship
the pam.d files and other auxilliary stuff in a separate tarball (which
will be shared with 3.4.x) to keep FILESDIR a bit cleaner.