x11-libs/cairo: Add explicit dep to media-libs/mesa-9.2.0-r1
x11-libs/cairo-1.12.16-r1 is depending against media-libs/mesa with
MULTILIB_USEDEP use dependencies. On portage the first version that
supports these dependencies is mesa-9.2.0-r1. Hence it makes sense to
add an explicit dependency for cairo to require 9.2.0-r1.

Without this patch portage wanted me to downgrade mesa to 9.0.3 and
giving me slot conflicts. With this patch, it prompts me to unmask
mesa-9.2.0-r1 and after that continues nicely.