gnome-base/librsvg: copy over stable keywords from older 2.45.92
for multilib to work properly with rust-bin, a full i686 rust-std is needed. There is now a hacky ebuild in the overlay that does provide this, it should be integrated into rust-bin from the main tree at some point in the future.
If you test this and run into problems, please open an issue on the gnome-next github page.