Add systemd patchset to gnome-settings-daemon
These patches (from upstream Git, no release tag as of my writing this)
set up systemd services for the various components of
GNOME-Settings-Daemon and ensure that they're started *after* XWayland
(if GNOME is riding atop Wayland and not an XOrg server).  This seems to
be necessary when starting a GNOME session with bleeding-edge
Mutter/GNOME-Shell on a systemd-managed Linux instance, and is
definitely neccessary when starting any GNOME session (including via
GDM) on the same system type when sys-apps/dbus is merged with the
'user-session' USE flag set.  Without full systemd service support,
gnome-settings-daemon components can freak out when they don't see
XWayland, and if they fail to start, the whole GNOME session fails.